refactor: 기능별 객체 분리 및 구조 개선
