+++ src/main/java/com/project/openmarket/domain/cash/dto/request/PaymentConfirmRequestDto.java
@@ -1,11 +1,11 @@
-package com.project.openmarket.domain.cache.dto.request;
+package com.project.openmarket.domain.cash.dto.request;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public record PaymentConfirmRequestDto(String orderId, String paymentKey, String amount) {
 
-	public Map<String, String> toMap(){
+	public Map<String, String> toMap() {
 		Map<String, String> map = new HashMap<>();
 		map.put("orderId", orderId);
 		map.put("paymentKey", paymentKey);
