+++ src/main/java/com/project/openmarket/domain/cash/dto/request/PaymentRequestDto.java
@@ -1,0 +1,11 @@
+package com.project.openmarket.domain.cash.dto.request;
+
+import com.project.openmarket.domain.cash.entity.CashHistory;
+import com.project.openmarket.domain.cash.entity.enums.CashStatus;
+import com.project.openmarket.domain.user.entity.Consumer;
+
+public record PaymentRequestDto(String orderId, String amount) {
+	public CashHistory toEntity(Consumer consumer) {
+		return new CashHistory(consumer, Long.parseLong(amount), orderId, null, CashStatus.PENDING);
+	}
+}
