+++ src/main/java/com/project/openmarket/domain/cash/dto/response/PaymentFailResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.openmarket.domain.cache.dto.response;
+package com.project.openmarket.domain.cash.dto.response;
 
 public record PaymentFailResponseDto(String code, String message, String orderId) {
 }
