+++ src/main/java/com/project/openmarket/domain/cash/entity/CashHistory.java
@@ -1,4 +1,4 @@
-package com.project.openmarket.domain.cache.entity;
+package com.project.openmarket.domain.cash.entity;
 
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
@@ -7,10 +7,9 @@ import org.springframework.data.mongodb.core.mapping.DocumentReference;
 import org.springframework.data.mongodb.core.mapping.Field;
 
 import com.project.openmarket.domain.base.entity.BaseTime;
-import com.project.openmarket.domain.cache.entity.enums.CashStatus;
-import com.project.openmarket.domain.cache.entity.enums.PayType;
+import com.project.openmarket.domain.cash.entity.enums.CashStatus;
+import com.project.openmarket.domain.cash.entity.enums.PayType;
 import com.project.openmarket.domain.user.entity.Consumer;
-import com.project.openmarket.domain.user.entity.User;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -53,11 +52,11 @@ public class CashHistory extends BaseTime {
 		this.paymentKey = paymentKey;
 	}
 
-	public CashHistory(Consumer consumer, Long amount, String orderId, String paymentKey, CashStatus status){
+	public CashHistory(Consumer consumer, Long amount, String orderId, String paymentKey, CashStatus status) {
 		this(consumer, amount, status, null, orderId, null, paymentKey);
 	}
 
-	public void updateCashStatus(CashStatus status){
+	public void updateCashStatus(CashStatus status) {
 		this.status = status;
 	}
 }
