+++ src/main/java/com/project/openmarket/domain/cash/entity/enums/PayType.java
@@ -1,4 +1,4 @@
-package com.project.openmarket.domain.cache.entity.enums;
+package com.project.openmarket.domain.cash.entity.enums;
 
 import lombok.Getter;
