+++ src/main/java/com/project/openmarket/domain/cash/repository/CashHistoryRepository.java
@@ -1,11 +1,11 @@
-package com.project.openmarket.domain.cache.repository;
+package com.project.openmarket.domain.cash.repository;
 
 import java.util.Optional;
 
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.project.openmarket.domain.cache.entity.CashHistory;
+import com.project.openmarket.domain.cash.entity.CashHistory;
 import com.project.openmarket.global.exception.CustomException;
 import com.project.openmarket.global.exception.enums.ExceptionConstants;
 
@@ -13,7 +13,7 @@ public interface CashHistoryRepository extends MongoRepository<CashHistory, Obje
 
 	Optional<CashHistory> findByOrderId(String orderId);
 
-	default CashHistory getByOrderId(String orderId){
+	default CashHistory getByOrderId(String orderId) {
 		return findByOrderId(orderId)
 			.orElseThrow(() -> new CustomException(ExceptionConstants.NOT_FOUND_ORDER));
 	}
